+++ kong/plugins/post-function/handler.lua
@@ -1,1 +1,1 @@
-return require("kong.plugins.pre-function._handler")("post-function", -1000)
+return require("kong.plugins.pre-function._handler")(-1000)