+++ kong/plugins/pre-function/_handler.lua
@@ -1,45 +1,131 @@
 -- handler file for both the pre-function and post-function plugin
-return function(plugin_name, priority)
-  local loadstring = loadstring
-  local insert = table.insert
-  local ipairs = ipairs
 
-  local config_cache = setmetatable({}, { __mode = "k" })
+
+local config_cache do
+
+
+  local no_op = function() end
+
+
+  -- compiles the array for a phase into a single function
+  local function compile_phase_array(phase_funcs)
+    if not phase_funcs or #phase_funcs == 0 then
+      -- nothing to do for this phase
+      return no_op
+    else
+      -- compile the functions we got
+      local compiled = {}
+      for i, func_string in ipairs(phase_funcs) do
+        local func = loadstring(func_string)
+
+        local first_run_complete = false
+        compiled[i] = function()
+          -- this is a temporary closure, that will replace itself
+          if not first_run_complete then
+            first_run_complete = true
+            local result = func() --> this might call ngx.exit()
+
+            -- if we ever get here, then there was NO early exit from a 0.1.0
+            -- type config
+            if type(result) == "function" then
+              -- this is a new function (0.2.0+), with upvalues
+              -- the first call to func above only initialized it, so run again
+              func = result
+              compiled[i] = func
+              func() --> this again, may do an early exit
+            end
+
+            -- if we ever get here, then there was no early exit from either
+            -- 0.1.0 or 0.2.0+ code
+            -- Replace the entry of this closure in the array with the actual
+            -- function, since the closure is no longer needed.
+            compiled[i] = func
+
+          else
+            -- first run is marked as complete, but we (this temporary closure)
+            -- are being called again. So we are here only if the initial
+            -- function call did an early exit.
+            -- So replace this closure now;
+            compiled[i] = func
+            -- And call it again, for this 2nd run;
+            func()
+          end
+          -- unreachable
+        end
+      end
+
+      -- now return a function that executes the entire array
+      return function()
+        for _, f in ipairs(compiled) do f() end
+      end
+    end
+  end
+
+
+  local phases = { "certificate", "rewrite", "access",
+                   "header_filter", "body_filter", "log",
+                   "functions" } -- <-- this one being legacy
+
+
+  config_cache = setmetatable({}, {
+    __mode = "k",
+    __index = function(self, config)
+      -- config was not found yet, so go and compile our config functions
+      local runtime_funcs = {}
+      for _, phase in ipairs(phases) do
+        local func = compile_phase_array(config[phase])
+
+        if phase == "functions" then
+          if func == no_op then
+            func = nil -- do not set a "functions" key, since we won't run it anyway
+          else
+            -- functions, which is legacy is specified, so inject as "access". The
+            -- schema already prevents "access" and "functions" to co-exist, so
+            -- this should be safe.
+            phase = "access"
+          end
+        end
+
+        runtime_funcs[phase] = func
+      end
+      -- store compiled results in cache, and return them
+      self[config] = runtime_funcs
+      return runtime_funcs
+    end
+  })
+end
+
+
+
+return function(priority)
 
   local ServerlessFunction = {
     PRIORITY = priority,
     VERSION = "0.3.1",
   }
 
+  function ServerlessFunction:certificate(config)
+    config_cache[config].certificate()
+  end
+
+  function ServerlessFunction:rewrite(config)
+    config_cache[config].rewrite()
+  end
+
   function ServerlessFunction:access(config)
+    config_cache[config].access()
+  end
 
-    local functions = config_cache[config]
+  function ServerlessFunction:header_filter(config)
+    config_cache[config].header_filter()
+  end
 
-    if not functions then
-      -- first call, go compile the functions
-      functions = {}
-      for _, fn_str in ipairs(config.functions) do
-        local func1 = loadstring(fn_str)    -- load it
-        local _, func2 = pcall(func1)       -- run it
-        if type(func2) ~= "function" then
-          -- old style (0.1.0), without upvalues
-          insert(functions, func1)
-        else
-          -- this is a new function (0.2.0+), with upvalues
-          insert(functions, func2)
+  function ServerlessFunction:body_filter(config)
+    config_cache[config].body_filter()
+  end
 
-          -- the first call to func1 above only initialized it, so run again
-          func2()
-        end
-      end
-
-      config_cache[config] = functions
-      return  -- must return since we allready executed them
-    end
-
-    for _, fn in ipairs(functions) do
-      fn()
-    end
+  function ServerlessFunction:log(config)
+    config_cache[config].log()
   end