+++ kong/plugins/pre-function/handler.lua
@@ -1,1 +1,1 @@
-return require("kong.plugins.pre-function._handler")("pre-function", math.huge)
+return require("kong.plugins.pre-function._handler")(math.huge)